backend/requirements.txt
docker-compose==1.29.2
docker==5.0.3
django==3.2.13
python-dotenv==0.20.0
djangorestframework==3.13.1
djoser==2.1.0
psycopg2-binary==2.9.3
django-filter==21.1
Pillow==9.1.0
drf-extra-fields==3.4.0
gunicorn==20.1.0
django-import-export==2.8.0
